Move `NewSession` extension trait into separate module
This is a purely internal refactor, for ease of code navigation.
